Move switch data back into the platform-file
